refactor(gateway): extract session hub types
