+++ code/apps/bpane-gateway/src/session_hub.rs
@@ -11,45 +11,13 @@ mod policy;
 mod pump;
 mod refresh;
 mod telemetry;
+mod types;
 
 pub use self::telemetry::SessionTelemetrySnapshot;
+pub use self::types::{ClientHandle, ResizeResult, SubscribeError};
 
 /// Broadcast channel capacity. At 30fps, 1024 frames is ~34 seconds of buffer.
 const BROADCAST_CAPACITY: usize = 1024;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum SubscribeError {
-    ViewerLimitReached { max_viewers: u32 },
-}
-
-impl std::fmt::Display for SubscribeError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::ViewerLimitReached { max_viewers } => {
-                write!(f, "viewer limit reached (max {max_viewers})")
-            }
-        }
-    }
-}
-
-impl std::error::Error for SubscribeError {}
-
-/// Handle returned to each connecting client.
-#[derive(Debug)]
-pub struct ClientHandle {
-    /// Broadcast receiver for host->client frames.
-    pub from_host: broadcast::Receiver<Arc<Frame>>,
-    /// Sender for client->host frames (cloned from hub).
-    pub to_host: mpsc::Sender<Frame>,
-    /// Unique client ID within this session.
-    pub client_id: u64,
-    /// Whether this client is the session owner (first to connect).
-    pub is_owner: bool,
-    /// Frames to send immediately on connect (cached SessionReady + last keyframe).
-    pub initial_frames: Vec<Arc<Frame>>,
-    /// The current locked resolution (only set for non-owner clients).
-    pub locked_resolution: Option<(u16, u16)>,
-}
 
 /// Central fan-out/merge coordinator for a single host agent session.
 ///
@@ -221,14 +189,6 @@ impl SessionHub {
     fn record_refresh_burst(&self, tiles_requested: u64) {
         telemetry::record_refresh_burst(self, tiles_requested);
     }
-}
-
-/// Result of a resize request.
-pub enum ResizeResult {
-    /// Request was forwarded to the host agent.
-    Applied,
-    /// Resolution is locked by the owner — use these dimensions.
-    Locked(u16, u16),
 }
 
 #[cfg(test)]
